feat: update default JSON formatter to Prettier, refactor repository state management, and enhance diff generation logic
